image-preview/worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.55",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]
